Moved all dialogs to new file. Added 2 new previews for select and deselect workout.
